Refactor Other CI repo onboarding into one file (#2807)
