+++ src/pages/RepoPage/CoverageOnboarding/OtherCI/TerminalInstructions/InstructionBox.spec.tsx
@@ -3,12 +3,12 @@ import userEvent from '@testing-library/user-event'
 
 import config from 'config'
 
-import { InstructionBoxOrgToken } from './InstructionBoxOrgToken'
+import { InstructionBox } from './InstructionBox'
 
 jest.mock('copy-to-clipboard', () => () => true)
 jest.mock('config')
 
-describe('InstructionBoxOrgToken', () => {
+describe('InstructionBox', () => {
   function setup({ isSelfHosted = false } = {}) {
     const user = userEvent.setup()
     config.IS_SELF_HOSTED = isSelfHosted
@@ -22,35 +22,35 @@ describe('InstructionBoxOrgToken', () => {
     beforeEach(() => setup())
 
     it('renders Linux button', () => {
-      render(<InstructionBoxOrgToken />)
+      render(<InstructionBox />)
 
       const button = screen.getByRole('button', { name: 'Linux' })
       expect(button).toBeInTheDocument()
     })
 
     it('renders MacOS button', () => {
-      render(<InstructionBoxOrgToken />)
+      render(<InstructionBox />)
 
       const button = screen.getByRole('button', { name: 'macOS' })
       expect(button).toBeInTheDocument()
     })
 
     it('renders Alpine Linux button', () => {
-      render(<InstructionBoxOrgToken />)
+      render(<InstructionBox />)
 
       const button = screen.getByRole('button', { name: 'Alpine Linux' })
       expect(button).toBeInTheDocument()
     })
 
     it('renders Windows button', () => {
-      render(<InstructionBoxOrgToken />)
+      render(<InstructionBox />)
 
       const button = screen.getByRole('button', { name: 'Windows' })
       expect(button).toBeInTheDocument()
     })
 
     it('renders linux default instructions', () => {
-      render(<InstructionBoxOrgToken />)
+      render(<InstructionBox />)
 
       const instruction = screen.queryByText(
         'curl -Os https://cli.codecov.io/latest/linux/codecov/'
@@ -62,7 +62,7 @@ describe('InstructionBoxOrgToken', () => {
   describe('when click on windows', () => {
     it('renders the windows instruction', async () => {
       const { user } = setup()
-      render(<InstructionBoxOrgToken />)
+      render(<InstructionBox />)
 
       const WindowsButton = screen.getByRole('button', { name: 'Windows' })
       await user.click(WindowsButton)
@@ -75,7 +75,7 @@ describe('InstructionBoxOrgToken', () => {
   describe('when click on Alpine', () => {
     it('renders the Alpine instruction', async () => {
       const { user } = setup()
-      render(<InstructionBoxOrgToken />)
+      render(<InstructionBox />)
 
       const alpineButton = screen.getByRole('button', { name: 'Alpine Linux' })
       await user.click(alpineButton)
@@ -90,7 +90,7 @@ describe('InstructionBoxOrgToken', () => {
   describe('when click on Mac', () => {
     it('renders the Mac instruction', async () => {
       const { user } = setup()
-      render(<InstructionBoxOrgToken />)
+      render(<InstructionBox />)
 
       const macButton = screen.getByRole('button', { name: 'macOS' })
       await user.click(macButton)
@@ -105,7 +105,7 @@ describe('InstructionBoxOrgToken', () => {
   describe('when user is a self hosted user', () => {
     it('renders windows specific instruction', async () => {
       const { user } = setup({ isSelfHosted: true })
-      render(<InstructionBoxOrgToken />)
+      render(<InstructionBox />)
 
       const WindowsButton = screen.getByRole('button', { name: 'Windows' })
       await user.click(WindowsButton)
