--- src/pages/RepoPage/CoverageOnboarding/OtherCI/TerminalInstructions/index.js
@@ -1,2 +1,0 @@
-export { InstructionBoxRepoToken } from './InstructionBoxRepoToken'
-export { InstructionBoxOrgToken } from './InstructionBoxOrgToken'
