+++ src/pages/RepoPage/CoverageOnboarding/OtherCI/TerminalInstructions/index.ts
@@ -1,0 +1,1 @@
+export { InstructionBox } from './InstructionBox'
